requirements.txt
langchain
chromadb
pypdf
pytest
langchain-community
flask-cors
langchainhub
langchain_core